Add files via upload
A& B are two different landing pages for hotel bookings company wants to know which page will perform better to drive bookings. 

Null Hypothesis: Both versions A and B have an equal probability of conversion or driving customer booking. In other words, there is no difference or no effect between the A and B versions

Alternative Hypothesis: Versions both A and B possess a different probability of conversion or driving customer booking and there is a difference between A and B versions. Version B is better than version A in driving customer bookings.

We will perform tests for two proportions - Z-test